Compact VGG-inspired ConvNet for CIFAR-10
Introducing YoshiNetV2, a custom-designed convolutional neural network for the CIFAR-10 dataset. Inspired by the VGG architecture, YoshiNetV2 is streamlined with fewer layers to balance performance and efficiency. The implementation in PyTorch leverages nn.Sequential for a clear and concise structure. This notebook covers the entire workflow, including data preprocessing, model training, and postprocessing analysis. Aiming for a balance between depth and simplicity, YoshiNetV2 is a step forward in personal ConvNet design.